pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ground-events"
version = "0.1.0"
description = "Multi-site event aggregation system for tracking events from various sources"
authors = [
    {name = "Steve Androulakis", email = "steve@example.com"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8.1"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests",
    "beautifulsoup4",
    "lxml",
    "aiohttp",
    "anthropic>=0.40.0",
    "temporalio>=1.9.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=3.4.8",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "aioresponses",
    "pytest-mock",
    "freezegun",
    "black",
    "isort",
    "flake8",
    "mypy",
]

[project.scripts]
ground-events = "around_the_grounds.main:main"

[project.urls]
Homepage = "https://github.com/steveandroulakis/ground-events"
Repository = "https://github.com/steveandroulakis/ground-events"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
# Exclude temporal test files due to Temporal SDK internal type issues
exclude = ["tests/temporal/test_workflows.py"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "slow: Slow tests that involve network calls"
]

[dependency-groups]
dev = [
    "aioresponses>=0.7.8",
    "autoflake>=2.3.1",
    "black>=24.8.0",
    "flake8>=7.1.2",
    "freezegun>=1.5.2",
    "isort>=5.13.2",
    "mypy>=1.14.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.1",
]
